08-workoutTracker/requirements.txt
requests
black
mypy
python-dotenv